# doc/api.yaml
openapi: 3.0.3
info: 
  title: WASAPhoto
  description: |- 
    This OpenAPI document describes the APIs required for the Project "WASAPhoto" 
    of the WASA exam of Sapienza University of Rome.
  version: 1.0.1
servers:
  - url: "http://localhost:3000"
...
paths:
...
  /users/{userID}/photos/{photoID}:
    parameters:
      - $ref: "#/components/parameters/userIDParam"
      - $ref: "#/components/parameters/photoIDParam"
    delete:
      security:
        - bearerAuthentication: []
      tags: ["photos"]
      summary: Delete a Photo uploaded on WASAPhoto
      description: |-
        Given the user id of the user that uploaded the photo and the photo id of the 
        photo itself, it deletes the photo from the photos uploaded by the user
      operationId: deletePhoto
      responses:
        "204":
          description: Photo successfully removed
        "400":
          $ref: "#/components/responses/400BadRequest" 
        "403":
          $ref: "#/components/responses/403Forbidden" 
        "404":
          $ref: "#/components/responses/404NotFound" 
        "500":
          $ref: "#/components/responses/500InternalServerError"   
  /users/{userID}/follow/{followID}:
    parameters:
      - $ref: "#/components/parameters/userIDParam"
      - $ref: "#/components/parameters/followIDParam"
    put:
      security:
        - bearerAuthentication: []
      tags: ["follow"]
      summary: Follow an existent user on WASAPhoto
      description: |-
        Given the id of the user who wants to follow another one (userID) and the
        id of the user who will be followed (followID), if they are actually 
        users of WASAPhoto, this method allows the follow action
      operationId: followUser
      responses:
        "204": 
          description: User successfully followed
        "400":
          $ref: "#/components/responses/400BadRequest" 
        "403":
          $ref: "#/components/responses/403Forbidden" 
        "404":
          $ref: "#/components/responses/404NotFound" 
        "500":
          $ref: "#/components/responses/500InternalServerError"      
          
    delete:
      security:
        - bearerAuthentication: []
      tags: ["follow"]
      summary: Unfollow an existent user on WASAPhoto
      description: |-
        Given the id of the user who wants to unfollow another one (userID) and the
        id of the user who will be unfollowed (followID), if they are actually 
        users of WASAPhoto and the user follows the other one, this method 
        allows the unfollow action
      operationId: unfollowUser
      responses:
        "204": 
          description: User successfully unfollowed
        "400":
          $ref: "#/components/responses/400BadRequest" 
        "403":
          $ref: "#/components/responses/403Forbidden" 
        "404":
          $ref: "#/components/responses/404NotFound" 
        "500":
          $ref: "#/components/responses/500InternalServerError"
...
  /users/{userID}/photos/{photoID}/likes/{likeID}:
    parameters:
      - $ref: "#/components/parameters/userIDParam"
      - $ref: "#/components/parameters/photoIDParam"
      - $ref: "#/components/parameters/likeIDParam"
    put:
      security:
        - bearerAuthentication: []
      tags: ["photos"]
      summary: Like a photo
      description: |-
        Given the photo id, and the id of the User who wants to add a 
        like to the Photo (likeID), it adds a like
      operationId: likePhoto
      responses:
        "201":
          description: Like succesfully added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Like" 
        "204": 
          description: Like already exists
        "400":
          $ref: "#/components/responses/400BadRequest"
        "403": 
          description: Users can't like their own photos
        "404":
          $ref: "#/components/responses/404NotFound" 
        "500":
          $ref: "#/components/responses/500InternalServerError"
          
    delete:
      security:
        - bearerAuthentication: []
      tags: ["photos"]
      summary: Unlike a photo
      description: |-
        Given the photo id and the id of the User who liked the Photo (likeID), deletes the Like with the user id 
        that corresponds to the one given
      operationId: unlikePhoto
      responses:
        "204":
          description: Removed successfully a like from the photo
        "400":
          $ref: "#/components/responses/400BadRequest"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound" 
        "500":
          $ref: "#/components/responses/500InternalServerError"
...
  /users/{userID}/photos/{photoID}/comments/{commentID}:
    parameters:
      - $ref: "#/components/parameters/userIDParam"
      - $ref: "#/components/parameters/photoIDParam"
      - $ref: "#/components/parameters/commentIDParam"
    delete:
      security:
        - bearerAuthentication: []
      tags: ["photos"]
      summary: Uncomment a Photo
      description: |- 
        Given the photo id and a comment id, deletes the Comment with the comment id 
        that corresponds to the one given
      operationId: uncommentPhoto
      responses:
        "204":
          description: Comment successfully removed
        "400":
          $ref: "#/components/responses/400BadRequest"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"          
        "500":
          $ref: "#/components/responses/500InternalServerError"
...
  /users/{userID}/ban/{banID}:
    parameters:
      - $ref: "#/components/parameters/userIDParam"
      - $ref: "#/components/parameters/banIDParam"
    put:
      security:
        - bearerAuthentication: []
      tags: ["ban"]
      summary: Ban a User
      description: |- 
        Given the id of the user who wants to ban another one (userID) and the
        id of the user to ban (banID), this method allows the ban operation, meaning
        that the banned user will no longer be able to visualize the user profile of
        the user who banned him
      operationId: banUser
      responses:
        "204": 
          description: "User successfully banned, he will no longer be able to visualize your profile"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound" 
        "500":
          $ref: "#/components/responses/500InternalServerError"

    delete:
      security:
        - bearerAuthentication: []
      tags: ["ban"]
      summary: Unban a User
      description: |- 
        Given the id of the user who wants to unban another one (userID) and the
        id of the user to unban (banID), this method allows the unban operation 
        (only if the user was previously in the ban status), meaning that the 
        ex-banned user will be able to re-visualize the user profile of 
        the user who previously banned him
      operationId: unbanUser
      responses:
        "204": 
          description: "User successfully unbanned, he will be able to re-visualize your profile"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "403":
          description: The user isn't banned
        "404":
          $ref: "#/components/responses/404NotFound"
        "500":
          $ref: "#/components/responses/500InternalServerError"
components:
  securitySchemes:
    bearerAuthentication:
      description: |-
        Bearer Authentication that uses the user identifier in place of the token
        that needs to be passed into the Authorization header
      type: http
      scheme: bearer
  parameters:
    userIDParam:
      name: userID
      description: The unique identifier of the User
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/User/properties/userID"
    followIDParam:
      name: followID
      description: Identifier of the user who will be followed/unfollowed
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/User/properties/userID"
    banIDParam:
      name: banID
      description: Identifier of the user who will be banned/unbanned
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/User/properties/userID"
    photoIDParam:
      name: photoID
      description: The unique identifier of the Photo
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Photo/properties/photoID"
    commentIDParam:
      name: commentID
      description: The unique identifier of the Comment
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Comment/properties/commentID"
    likeIDParam:
      name: likeID
      description: The unique identifier of the User who added the Like
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/User/properties/userID"
...
  responses:
    400BadRequest:
      description: The request is invalid or incorrect
    403Forbidden:
      description: Access to the requested resource is forbidden. The authenticated user lacks the necessary permissions
    404NotFound:
      description: The requested resource was not found
    500InternalServerError:
      description: An internal server error occured
  schemas:
...
    Like:
      description: This object represents a single Like that a User can add onto a Photo
        on WASAPhoto
      type: object
      properties:
        userID:
          description: The unique identifier of the User who added the Like
          type: integer
          example: 202
          readOnly: true
        photoID:
          description: The unique identifier of the Photo target
          type: integer
          example: 567
          readOnly: true 
        date:
          description: Date that represents when the Like was added
          type: string
          format: date-time
          example: "2023-11-02T16:50:30Z"
          minLength: 20
          maxLength: 20